refactor: server wrappers and snapshot handling
