+++ types/copilotkit.d.ts
@@ -1,0 +1,6 @@
+declare module '@copilotkit/react-ui' {
+  export const CopilotKit: any;
+  export const useCopilotReadable: any;
+  export const CopilotSidebar: any;
+  export const CopilotChat: any;
+}
