+++ types/storybook-theming.d.ts
@@ -1,0 +1,3 @@
+declare module '@storybook/theming' {
+  export function create(options: any): any
+}
